# (Feat) 273% improvement GCS Bucket Logger - use Batched Logging