--- a/litellm/proxy/proxy_config.yaml
+++ b/litellm/proxy/proxy_config.yaml
@@ -7,10 +7,4 @@ model_list:
 
 
 litellm_settings:
-  callbacks: ["prometheus"]
-  service_callback: ["prometheus_system"]
-
-
-general_settings:
-  allow_requests_on_db_unavailable: true
-
+  callbacks: ["gcs_bucket"]