## Improve GCS Bucket scalability 

| Metric | Before 🔴 | After 🟢 | Improvement | % Improvement |
|---------|------------|-----------|--------------|----------------|
| Median latency with LiteLLM Proxy  | 1500ms | 100ms | 15x better | 93.3% |
| RPS | 30 | 112 | 3.73x better | 273% |


<!-- e.g. "Implement user authentication feature" -->

## Relevant issues

<!-- e.g. "Fixes #000" -->

## Type

<!-- Select the type of Pull Request -->
<!-- Keep only the necessary ones -->

🆕 New Feature


## Changes

<!-- List of changes -->

## [REQUIRED] Testing - Attach a screenshot of any new tests passing locall
If UI changes, send a screenshot/GIF of working UI fixes

<!-- Test procedure -->